Enable Groovy Gradle dependency edit tests
