+++ crates/core/src/codemod_steps/utils/mod.rs
@@ -1,3 +1,4 @@
 pub(crate) mod ast;
+pub(crate) mod gradle;
 pub(crate) mod ranges;
 pub(crate) mod xml;
